pill_reader/backend/requirements.txt
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.4.2
pyenchant==1.6.8
pytesseract==0.1.6
requests==2.11.1
Werkzeug==0.11.11
